+++ src/DSH-Desktop/UninstallProgram.cs
@@ -14,8 +14,8 @@ using Microsoft.Win32;
 [assembly: AssemblyCompany("DSH Desktop contributors")]
 [assembly: AssemblyProduct("DSH Desktop")]
 [assembly: AssemblyCopyright("Copyright (c) 2026 DSH Desktop contributors")]
-[assembly: AssemblyVersion("0.4.4.0")]
-[assembly: AssemblyFileVersion("0.4.4.0")]
+[assembly: AssemblyVersion("0.4.5.0")]
+[assembly: AssemblyFileVersion("0.4.5.0")]
 
 internal sealed class UninstallConfiguration
 {